Improve resubscription logic.
Drop the use of EventTypes, which is deprecated in Redfish.
The HPE servers were also doing something odd, where they initially
report the EventTypes in the subscription, but later do not. This
made it look like the matching subscription disappeared, so dftfd
was subscribing again, resulting in duplicate subscriptions and
duplicate data.

Also simplify the subscription matching logic to only look at
the combination of Destination and Context, which should allow
us to also purge some duplicates should they arise again in the
future.